Add get_deals_pipeline_by_id
Add `get_deals_pipeline_by_id` to `crm_pipelines`, allowing to retrieve a deal from its id.
